Extract lambda-natchez module
This is to allow consumers the choice of natchez
or otel4s

No breaking changes except requiring a
`feral.lambda.natchez._` import
