+++ lambda-natchez/shared/src/main/scala/feral/lambda/natchez/AwsTags.scala
@@ -14,16 +14,12 @@
  * limitations under the License.
  */
 
-package feral.lambda
+package feral.lambda.natchez
 
-import natchez.Kernel
+import natchez.TraceValue
 
-trait KernelSource[Event] {
-  def extract(event: Event): Kernel
-}
-
-object KernelSource {
-  @inline def apply[E](implicit ev: KernelSource[E]): ev.type = ev
-
-  def emptyKernelSource[E]: KernelSource[E] = _ => Kernel(Map.empty)
+object AwsTags {
+  private[this] val prefix = "aws"
+  def arn(s: String): (String, TraceValue) = s"$prefix.arn" -> s
+  def requestId(s: String): (String, TraceValue) = s"$prefix.requestId" -> s
 }
